Physics finally finished. I haven't run into any glitchy mechanics yet. Please notify me if you do.
I changed the font of the hud up a little, and also changed the colors of the font up a bit. 

The game now loads a different background for level 1 and other levels. 

Work In Progress: I am trying to not have the ball automatically start when the player dies, and instead have the player press Enter like at the beginning of the game. So, for now, I have a disconnectTimer function which can be activated by pressing T in the middle of the game. Unfortunately, since activating the timer requires the timer to not be active, and for some reason, disconnecting the timer doesn't make it inactive, you cannot reactivate the timer once you disconnect it.