Level Three added.
A bunch of stuff is in this update, but none of them completely work, therefore most of them have been commented out. Form class is the same as the last revision i.e. r92. To see what level 3 looks like, simply change levelNumber to 3. 

Powerups: They are acting weird, so I have commented them out for now, but this has to be the final version of the ball class, i.e. the changes that I have made to how the ball class works are crucial for the powerups to work properly. The timer also has to be set at 1, so thats another important thing. 

Score: I am working on it, its my first messing around with signals and slots. I made Board static temporarily, and i think that's included in this update. I don't think I need it static anymore. We'll see...