Powerups done sort of: There are now powerups and powerdowns in effect beginning from level two to level five. There is a speeding up of the ball and slowing down of the ball in effect as of now.
I tried implementing multiple balls, but for whatever reason, adding new balls through the ball class to the scene does not work the way it should, and instead messes up. 

Other bugs encountered: Once in a while, randomly, the ball will bounce off a block which might have a powerup, but the animation of the powerup coming down will never be shown and instead, the ball will automatically either slow down or speed up.