Storyline finished.
I will add a skip intro button which takes the user directly to levelOne, and also the background music playing during the story screen display. 

Another thing I noticed was that the score seems to be working when transitioning between levels. So if the score is 45 at the end of level one,it wont display that immediately, and instead does it when the next level is displayed.